Change constraints to shoot for 250MHz
This is the most the PS-PL AXI interfaces can run at,
so it's the upper maximum we could run at and is
as good of an an initial target as any.

Definitely possible to hit, but it will likely be very difficult.
We may be forced to lower this in the future.